requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
pdf2image
google-genai
python-dotenv
Pillow